Fully fix the Win64 build (test might still fail).

+++ source/eventcore/internal/utils.d
@@ -174,7 +174,7 @@ struct AlgebraicChoppedVector(TCommon, TSpecific...)
 
 /** Efficient bit set of dynamic size.
 */
-struct SmallIntegerSet(V : uint)
+struct SmallIntegerSet(V : size_t)
 {
 	private {
 		uint[][4] m_bits;
